docs(button-toggle): add playground example
